requirements.txt
openai>=1.0.0
pydantic>=2.0.0
pydantic-settings
wikipedia
tenacity
python-dotenv